Refactored operation interface and AbstractOperation class to account for architectural changes (Templates do not provide parameters)
